+++ app/client/js/electionSaver.js
@@ -65,7 +65,8 @@ function handleDbData(dbJson) {
 
 function handleZipCode(zip) {  
     $.ajax({
-        url: 'http://localhost:5000/' + zip,
+        // url: 'https://pog-the-vote.ue.r.appspot.com/' + zip,
+        url: 'http://localhost:8080/' + zip,
         type: 'GET',
         dataType: 'json',
         success: function(data){ 
@@ -74,8 +75,8 @@ function handleZipCode(zip) {
             $(".zip-text").text(zip);
 
             try {
-                electionOfficeData = data.parent_city.parent_county.election_office;
-                handleDbData(electionOfficeData);
+                // electionOfficeData = data.parent_city.parent_county.election_office;
+                handleDbData(data);
             } catch (error) {
                 console.error(error);
             }
@@ -88,16 +89,26 @@ function handleZipCode(zip) {
 
 $(document).ready(function() {
     $("input").on('keydown', function(evt) {
-        let numericOnly = $(this).val().replace(/[^0-9\.]/g,'');
-        let value = $(this).val() + String.fromCharCode(evt.which);
+        let char = String.fromCharCode(evt.which);
+        let numericOnly = char.replace(/[^0-9\.]/g,'');
+        let value = $(this).val() + char;
+
+        let thisInputId = $(this)[0].id.split('zip')[1]
+        let nextInput = $(`#zip${parseInt(thisInputId) + 1}`);
         if (value.length == 1 && evt.which != 9) { //9 is the tab character
-            let thisInputId = $(this)[0].id.split('zip')[1]
-            let nextInput = $(`#zip${parseInt(thisInputId) + 1}`);
             setTimeout(function() {
-                nextInput.focus();
-            }, 100);
+                if (numericOnly !== '') {
+                    nextInput.focus();
+                }
+            }, 50);
         }
-        if (value.length > 1) {
+        if (evt.which == 9) {
+            nextInput.focus();
+        }
+        if (numericOnly === ''&& evt.which != 37 && evt.which != 39 && evt.which != 8) {
+            return false;
+        }
+        if (value.length > 1 && evt.which != 37 && evt.which != 39 && evt.which != 8) {
             return false;
         }
     });
